Issue #110: VirtualAlloc() in AllocateAndFillBufferForTarget() fails for large values of RequestCount and BlockSize
. Modified the allocation strategy to only allocate the buffers that are needed and to allocated each buffer as a separate VirtualAlloc().  This seems to have fixed the problem (for now).

Note: This will waste some memory for allocations < 4k, but I think it's important for performance that each allocation is on a 4k boundary.  If the time comes when someone tries to test a large number of Requests with a small BlockSize this may have to be change to a strategy that uses malloc() for small allocations and VirtualAlloc() for large allocations.